Create Data Dig
Here you can see where I searched for only Fox studio with a ranking above 90. First you can see I put OR so it gave me different studios and not only Fox so I had to change that AND.  I also looked for movies in the year 2013 and above with the studio of SONY. 
I name a row top_ranked and only put movies that had a higher ranking than 95.
I wanted to see how many movies had a rank of 90 or 80 etc.. So I separated them by ranking with A, B, C, D, F, and made it show me the count as well next to it.